+++ CHANGES.md
@@ -3,6 +3,7 @@
 * **[Breaking change]** `sq.pareto` previously sampled from a Lomax distribution due to a complication with numpy. Now it properly samples from a Pareto distribution.
 * **[Breaking change]** lclip / rclip have been removed from triangular distribution because that doesn't make sense.
 * **[Breaking change]** You now can nest mixture and discrete distributions within mixture distributions.
+* **[Breaking change]** `sq.kelly` now raises an error if you put in a price below the market price. You can pass `error=False` to disable this and return to the old behavior.
 * Added `pert` distribution.
 * Added `sharpe_ratio` to utilities.